# Faster geometry conversion