--- a/cpp/open3d/t/geometry/PointCloud.cpp
+++ b/cpp/open3d/t/geometry/PointCloud.cpp
@@ -39,23 +39,22 @@ namespace open3d {
 namespace t {
 namespace geometry {
 
-PointCloud::PointCloud(core::Dtype dtype, const core::Device &device)
+PointCloud::PointCloud(const core::Device &device)
     : Geometry(Geometry::GeometryType::PointCloud, 3),
       device_(device),
       point_attr_(TensorMap("points")) {
-    SetPoints(core::Tensor::Zeros({0, 3}, dtype, device_));
+    ;
 }
 
 PointCloud::PointCloud(const core::Tensor &points)
-    : PointCloud(points.GetDtype(), points.GetDevice()) {
+    : PointCloud(points.GetDevice()) {
     points.AssertShapeCompatible({utility::nullopt, 3});
     SetPoints(points);
 }
 
 PointCloud::PointCloud(const std::unordered_map<std::string, core::Tensor>
                                &map_keys_to_tensors)
-    : PointCloud(map_keys_to_tensors.at("points").GetDtype(),
-                 map_keys_to_tensors.at("points").GetDevice()) {
+    : PointCloud(map_keys_to_tensors.at("points").GetDevice()) {
     if (map_keys_to_tensors.count("points") == 0) {
         utility::LogError("\"points\" attribute must be specified.");
     }
@@ -104,14 +103,12 @@ geometry::PointCloud PointCloud::FromLegacyPointCloud(
         const open3d::geometry::PointCloud &pcd_legacy,
         core::Dtype dtype,
         const core::Device &device) {
-    geometry::PointCloud pcd(dtype, device);
+    geometry::PointCloud pcd(device);
     if (pcd_legacy.HasPoints()) {
         pcd.SetPoints(core::eigen_converter::EigenVector3dVectorToTensor(
                 pcd_legacy.points_, dtype, device));
     } else {
-        utility::LogWarning(
-                "Creating from an empty legacy pointcloud, an empty pointcloud "
-                "with default dtype and device will be created.");
+        utility::LogWarning("Creating from an empty legacy PointCloud.");
     }
     if (pcd_legacy.HasColors()) {
         pcd.SetPointColors(core::eigen_converter::EigenVector3dVectorToTensor(
@@ -127,28 +124,17 @@ geometry::PointCloud PointCloud::FromLegacyPointCloud(
 open3d::geometry::PointCloud PointCloud::ToLegacyPointCloud() const {
     open3d::geometry::PointCloud pcd_legacy;
     if (HasPoints()) {
-        const core::Tensor &points = GetPoints();
-        pcd_legacy.points_.reserve(points.GetLength());
-        for (int64_t i = 0; i < points.GetLength(); i++) {
-            pcd_legacy.points_.push_back(
-                    core::eigen_converter::TensorToEigenVector3d(points[i]));
-        }
+        pcd_legacy.points_ =
+                core::eigen_converter::TensorToEigenVector3dVector(GetPoints());
     }
     if (HasPointColors()) {
-        const core::Tensor &colors = GetPointColors();
-        pcd_legacy.colors_.reserve(colors.GetLength());
-        for (int64_t i = 0; i < colors.GetLength(); i++) {
-            pcd_legacy.colors_.push_back(
-                    core::eigen_converter::TensorToEigenVector3d(colors[i]));
-        }
+        pcd_legacy.colors_ = core::eigen_converter::TensorToEigenVector3dVector(
+                GetPointColors());
     }
     if (HasPointNormals()) {
-        const core::Tensor &normals = GetPointNormals();
-        pcd_legacy.normals_.reserve(normals.GetLength());
-        for (int64_t i = 0; i < normals.GetLength(); i++) {
-            pcd_legacy.normals_.push_back(
-                    core::eigen_converter::TensorToEigenVector3d(normals[i]));
-        }
+        pcd_legacy.normals_ =
+                core::eigen_converter::TensorToEigenVector3dVector(
+                        GetPointNormals());
     }
     return pcd_legacy;
 }