--- a/cpp/open3d/t/geometry/TriangleMesh.cpp
+++ b/cpp/open3d/t/geometry/TriangleMesh.cpp
@@ -38,20 +38,15 @@ namespace open3d {
 namespace t {
 namespace geometry {
 
-TriangleMesh::TriangleMesh(core::Dtype vertex_dtype,
-                           core::Dtype triangle_dtype,
-                           const core::Device &device)
+TriangleMesh::TriangleMesh(const core::Device &device)
     : Geometry(Geometry::GeometryType::TriangleMesh, 3),
       device_(device),
       vertex_attr_(TensorMap("vertices")),
-      triangle_attr_(TensorMap("triangles")) {
-    SetVertices(core::Tensor({0, 3}, vertex_dtype, device_));
-    SetTriangles(core::Tensor({0, 3}, triangle_dtype, device_));
-}
+      triangle_attr_(TensorMap("triangles")) {}
 
 TriangleMesh::TriangleMesh(const core::Tensor &vertices,
                            const core::Tensor &triangles)
-    : TriangleMesh(vertices.GetDtype(), triangles.GetDtype(), [&]() {
+    : TriangleMesh([&]() {
           if (vertices.GetDevice() != triangles.GetDevice()) {
               utility::LogError(
                       "vertices' device {} does not match triangles' device "
@@ -63,6 +58,80 @@ TriangleMesh::TriangleMesh(const core::Tensor &vertices,
       }()) {
     SetVertices(vertices);
     SetTriangles(triangles);
+}
+
+geometry::TriangleMesh TriangleMesh::FromLegacyTriangleMesh(
+        const open3d::geometry::TriangleMesh &mesh_legacy,
+        core::Dtype float_dtype,
+        core::Dtype int_dtype,
+        const core::Device &device) {
+    if (float_dtype != core::Dtype::Float32 &&
+        float_dtype != core::Dtype::Float64) {
+        utility::LogError("float_dtype must be Float32 or Float64, but got {}.",
+                          float_dtype.ToString());
+    }
+    if (int_dtype != core::Dtype::Int32 && int_dtype != core::Dtype::Int64) {
+        utility::LogError("int_dtype must be Int32 or Int64, but got {}.",
+                          int_dtype.ToString());
+    }
+
+    TriangleMesh mesh(device);
+    if (mesh_legacy.HasVertices()) {
+        mesh.SetVertices(core::eigen_converter::EigenVector3dVectorToTensor(
+                mesh_legacy.vertices_, float_dtype, device));
+    } else {
+        utility::LogWarning("Creating from empty legacy TriangleMesh.");
+    }
+    if (mesh_legacy.HasVertexColors()) {
+        mesh.SetVertexColors(core::eigen_converter::EigenVector3dVectorToTensor(
+                mesh_legacy.vertex_colors_, float_dtype, device));
+    }
+    if (mesh_legacy.HasVertexNormals()) {
+        mesh.SetVertexNormals(
+                core::eigen_converter::EigenVector3dVectorToTensor(
+                        mesh_legacy.vertex_normals_, float_dtype, device));
+    }
+    if (mesh_legacy.HasTriangles()) {
+        mesh.SetTriangles(core::eigen_converter::EigenVector3iVectorToTensor(
+                mesh_legacy.triangles_, int_dtype, device));
+    }
+    if (mesh_legacy.HasTriangleNormals()) {
+        mesh.SetTriangleNormals(
+                core::eigen_converter::EigenVector3dVectorToTensor(
+                        mesh_legacy.triangle_normals_, float_dtype, device));
+    }
+    return mesh;
+}
+
+open3d::geometry::TriangleMesh TriangleMesh::ToLegacyTriangleMesh() const {
+    open3d::geometry::TriangleMesh mesh_legacy;
+    if (HasVertices()) {
+        mesh_legacy.vertices_ =
+                core::eigen_converter::TensorToEigenVector3dVector(
+                        GetVertices());
+    }
+    if (HasVertexColors()) {
+        mesh_legacy.vertex_colors_ =
+                core::eigen_converter::TensorToEigenVector3dVector(
+                        GetVertexColors());
+    }
+    if (HasVertexNormals()) {
+        mesh_legacy.vertex_normals_ =
+                core::eigen_converter::TensorToEigenVector3dVector(
+                        GetVertexNormals());
+    }
+    if (HasTriangles()) {
+        mesh_legacy.triangles_ =
+                core::eigen_converter::TensorToEigenVector3iVector(
+                        GetTriangles());
+    }
+    if (HasTriangleNormals()) {
+        mesh_legacy.triangle_normals_ =
+                core::eigen_converter::TensorToEigenVector3dVector(
+                        GetTriangleNormals());
+    }
+
+    return mesh_legacy;
 }
 
 }  // namespace geometry