--- a/cpp/tests/t/geometry/PointCloud.cpp
+++ b/cpp/tests/t/geometry/PointCloud.cpp
@@ -58,9 +58,8 @@ TEST_P(PointCloudPermuteDevices, DefaultConstructor) {
     EXPECT_FALSE(pcd.HasPointColors());
     EXPECT_FALSE(pcd.HasPointNormals());
 
-    // Default dtypes.
-    EXPECT_EQ(pcd.GetPoints().GetDevice(), core::Device("CPU:0"));
-    EXPECT_EQ(pcd.GetPoints().GetDtype(), core::Dtype::Float32);
+    // Default device.
+    EXPECT_EQ(pcd.GetDevice(), core::Device("CPU:0"));
 }
 
 TEST_P(PointCloudPermuteDevices, ConstructFromPoints) {
@@ -102,11 +101,11 @@ TEST_P(PointCloudPermuteDevices, ConstructFromPointDict) {
 
 TEST_P(PointCloudPermuteDevices, GetMinBound_GetMaxBound_GetCenter) {
     core::Device device = GetParam();
-    t::geometry::PointCloud pcd(core::Dtype::Float32, device);
+    t::geometry::PointCloud pcd(device);
 
-    core::Tensor& points = pcd.GetPointAttr("points");
-    points = core::Tensor(std::vector<float>{1, 2, 3, 4, 5, 6}, {2, 3},
-                          core::Dtype::Float32, device);
+    core::Tensor points = core::Tensor(std::vector<float>{1, 2, 3, 4, 5, 6},
+                                       {2, 3}, core::Dtype::Float32, device);
+    pcd.SetPoints(points);
 
     EXPECT_FALSE(pcd.IsEmpty());
     EXPECT_TRUE(pcd.HasPoints());
@@ -121,7 +120,7 @@ TEST_P(PointCloudPermuteDevices, GetMinBound_GetMaxBound_GetCenter) {
 TEST_P(PointCloudPermuteDevices, Translate) {
     core::Device device = GetParam();
 
-    t::geometry::PointCloud pcd(core::Dtype::Float32, device);
+    t::geometry::PointCloud pcd(device);
     core::Tensor translation(std::vector<float>{10, 20, 30}, {3},
                              core::Dtype::Float32, device);
 
@@ -142,10 +141,11 @@ TEST_P(PointCloudPermuteDevices, Translate) {
 
 TEST_P(PointCloudPermuteDevices, Scale) {
     core::Device device = GetParam();
-    t::geometry::PointCloud pcd(core::Dtype::Float32, device);
-    core::Tensor& points = pcd.GetPointAttr("points");
-    points = core::Tensor(std::vector<float>{0, 0, 0, 1, 1, 1, 2, 2, 2}, {3, 3},
-                          core::Dtype::Float32, device);
+    t::geometry::PointCloud pcd(device);
+    core::Tensor points =
+            core::Tensor(std::vector<float>{0, 0, 0, 1, 1, 1, 2, 2, 2}, {3, 3},
+                         core::Dtype::Float32, device);
+    pcd.SetPoints(points);
     core::Tensor center(std::vector<float>{1, 1, 1}, {3}, core::Dtype::Float32,
                         device);
     float scale = 4;
@@ -245,7 +245,7 @@ TEST_P(PointCloudPermuteDevices, Setters) {
     core::Tensor colors = core::Tensor::Ones({2, 3}, dtype, device) * 2;
     core::Tensor labels = core::Tensor::Ones({2, 3}, dtype, device) * 3;
 
-    t::geometry::PointCloud pcd(dtype, device);
+    t::geometry::PointCloud pcd(device);
 
     pcd.SetPoints(points);
     pcd.SetPointColors(colors);
@@ -279,7 +279,7 @@ TEST_P(PointCloudPermuteDevices, Has) {
     core::Device device = GetParam();
     core::Dtype dtype = core::Dtype::Float32;
 
-    t::geometry::PointCloud pcd(dtype, device);
+    t::geometry::PointCloud pcd(device);
     EXPECT_FALSE(pcd.HasPoints());
     EXPECT_FALSE(pcd.HasPointColors());
     EXPECT_FALSE(pcd.HasPointAttr("labels"));