--- a/cpp/tests/t/io/PointCloudIO.cpp
+++ b/cpp/tests/t/io/PointCloudIO.cpp
@@ -85,7 +85,7 @@ TEST_P(ReadWriteTPC, Basic) {
     ReadWritePCArgs args = GetParam();
     core::Device device("CPU", 0);
     core::Dtype dtype = core::Dtype::Float64;
-    t::geometry::PointCloud pc1(dtype, device);
+    t::geometry::PointCloud pc1(device);
 
     for (const auto &attr_tensor : pc_data_1) {
         const auto &attr = attr_tensor.first;
@@ -99,7 +99,7 @@ TEST_P(ReadWriteTPC, Basic) {
     EXPECT_TRUE(t::io::WritePointCloud(
             args.filename, pc1,
             {bool(args.write_ascii), bool(args.compressed), true}));
-    t::geometry::PointCloud pc2(dtype, device);
+    t::geometry::PointCloud pc2(device);
     EXPECT_TRUE(t::io::ReadPointCloud(args.filename, pc2,
                                       {"auto", false, false, true}));
 
@@ -115,7 +115,7 @@ TEST_P(ReadWriteTPC, Basic) {
     EXPECT_TRUE(t::io::WritePointCloud(
             args.filename, pc2,
             {bool(args.write_ascii), bool(args.compressed), true}));
-    t::geometry::PointCloud pc3(dtype, device);
+    t::geometry::PointCloud pc3(device);
     EXPECT_TRUE(t::io::ReadPointCloud(args.filename, pc3,
                                       {"auto", false, false, true}));
     for (const auto &attribute_rel_tol : args.attributes_rel_tols) {
@@ -130,7 +130,7 @@ TEST_P(ReadWriteTPC, WriteBadData) {
     ReadWritePCArgs args = GetParam();
     core::Device device("CPU", 0);
     core::Dtype dtype = core::Dtype::Float64;
-    t::geometry::PointCloud pc1(dtype, device);
+    t::geometry::PointCloud pc1(device);
 
     for (const auto &attr_tensor : pc_data_bad) {
         const auto &attr = attr_tensor.first;