… cpp/tests/t/geometry/TriangleMesh.cpp