- Add mesh conversion functions
- Constructor of pointcloud and trianglemesh no longer takes dtypes
- Several orders of magnitudes faster conversion to legacy `PointCloud` and `TriangleMesh`.

```
# Before: 1,000 points, 1,000 colors
------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations
------------------------------------------------------------------
ToLegacyPointCloud/CPU        21.5 ms         21.5 ms           30
ToLegacyPointCloud/CUDA       4470 ms         4470 ms            1


# After: 1,000 points, 1,000 colors
------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations
------------------------------------------------------------------
ToLegacyPointCloud/CPU       0.030 ms        0.030 ms        20128
ToLegacyPointCloud/CUDA       4.47 ms         4.47 ms          152


# After: 1,000,000 points, 1,000,000 colors
------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations
------------------------------------------------------------------
ToLegacyPointCloud/CPU        31.9 ms         30.0 ms           26
ToLegacyPointCloud/CUDA       23.3 ms         23.3 ms           30
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/intel-isl/open3d/2719)
<!-- Reviewable:end -->